% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley}
\alias{barley}
\title{Effect of Nitrolime and Superphosphate on Barley Yields}
\format{
\subsection{\code{barley}}{

A data frame with 20 rows and 4 columns:
\describe{
\item{Rep}{Factor. The replicate number}
\item{Nitrolime}{Factor. The nitrolime application rate.}
\item{Superphosphate}{Factor. The superphosphate application rate.}
\item{Yield}{Numeric. The yield in kg/ha.}
}
}
}
\source{
Saville, D. J., & Wood, G. R. (1991). Statistical Methods: The Geometric Approach. Springer Science & Business Media.
}
\usage{
barley
}
\description{
Data on barley yield for use in interaction contrast example.
}
\keyword{datasets}
