% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataClasses}
\alias{dataClasses}
\title{Extract variable classes}
\usage{
dataClasses(model)
}
\arguments{
\item{model}{}
}
\value{
a \code{\link{data.frame}}
}
\description{
Extract variable classes
}
\keyword{internal}
