% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exclusive_terms}
\alias{exclusive_terms}
\title{Identify Exclusive terms used to identify levels of random slope and intercept variables}
\usage{
exclusive_terms(model, term)
}
\arguments{
\item{model}{A \code{\link{glmmTMB}} model}

\item{term}{A model term}
}
\value{
a \code{\link{data.frame}}
}
\description{
Identify Exclusive terms used to identify levels of random slope and intercept variables
}
\keyword{internal}
