% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greenfeeds}
\alias{greenfeeds}
\title{Fodder Beet Yields}
\format{
\subsection{\code{greenfeeds}}{

A data frame with 32 rows and 3 columns:
\describe{
\item{Treatment}{Factor. Beet variety.}
\item{Rep}{Factor. The replicate number.}
\item{Yield}{Numeric. Beet yield in tons/ha}
...
}
}
}
\source{
Saville, D. J., & Wood, G. R. (1991). Statistical Methods: The Geometric Approach. Springer Science & Business Media.
}
\usage{
greenfeeds
}
\description{
Data on fodder beet yields for use in class contrasts example.
}
\keyword{datasets}
