% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{power_contrast}
\alias{power_contrast}
\title{Power of Contrasts Performed on Models Fit with glmmTMB}
\usage{
power_contrast(emm, contr_list, ddf = NULL, alpha = 0.05, n_sims = 10000, ...)
}
\arguments{
\item{emm}{An emmGrid object associated with a \code{glmmTMB} model.}

\item{contr_list}{A (named) list of contrast specifications.}

\item{ddf}{Either a method for calculating denominator degrees of freedom (
currently supported options are "df.residual", "asymptotic" and
"kenward-roger"), a numeric value, or NULL (the default), in which case a
method is selected based on the model type.}

\item{alpha}{Numeric. The nominal type I error rate. Defaults to 0.05.}

\item{n_sims}{Numeric. The number of simulations to use for calculating
type M error rate. If set to zero, the closed form, asymptotic
calculations are used.}

\item{...}{Other arguments passed to emmeans::contrast.}
}
\value{
A \code{powertable} object.
}
\description{
This function calculates the power of contrasts.
}
\examples{

library(emmeans)

# Create synthetic data set:
nrep = 8 # Number of replicates
nfac = 2 # Number of treatments
dat = expand.grid(Rep = factor(1:nrep),
                  Trt = factor(LETTERS[1:nfac])) 
dat$Y = ifelse(dat$Trt == 'A', 0, 3)

# "Fit" the model while setting residual standard deviation to 2:
mod = set_glmm(Y ~ Trt, data = dat, disp = 2)

# Create emmeans object:
emm = emmeans(mod, ~ Trt)

# Define contrasts (here there is only one):
contr = list('A-B' = c(1, -1))

# Calculate power of the contrast
power_contrast(emm, contr)

}
