% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{power_ftest}
\alias{power_ftest}
\title{Statistical Power of F-tests Performed on Models Fit with glmmTMB}
\usage{
power_ftest(mod, ddf = NULL, alpha = 0.05, ...)
}
\arguments{
\item{mod}{A \code{\link{glmmTMB}} model}

\item{ddf}{Either a method for calculating denominator degrees of freedom (
currently supported options are "df.residual", "asymptotic" and
"kenward-roger"), a numeric vector, or NULL (the default), in which case a
method is selected based on the model type.}

\item{alpha}{The nominal type I error rate. Defaults to 0.05.}

\item{...}{Other values passed to emmeans.}
}
\value{
A \code{powertable} object.
}
\description{
This function calculates the power of F-tests for each fixed
effect term in the model using one of several methods for determining the
denominator degrees of freedom.
}
\examples{

# Create synthetic data set:
nrep = 8 # Number of replicates
nfac = 2 # Number of treatments
dat = expand.grid(Rep = factor(1:nrep),
                  Trt = factor(LETTERS[1:nfac])) 
dat$Y = ifelse(dat$Trt == 'A', 0, 3)

# "Fit" the model while setting residual standard deviation to 2:
mod = set_glmm(Y ~ Trt, data = dat, disp = 2)

# Calculate power
power_ftest(mod)

}
