% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_method.R
\name{anova.glmmTMB}
\alias{anova.glmmTMB}
\title{Performs F-test on coefficients of a model fit via glmmTMB using one
of several available methods for calculating denomenator degrees of freedom}
\usage{
\method{anova}{glmmTMB}(model, method = "containment", type = 2, contr_sum = TRUE)
}
\arguments{
\item{model}{a \code{\link{glmmTMB}} model object}

\item{method}{Currently supported options are: "containment" (the default),
"pinheiro-bates", "asymptotic" and "satterthwaite".}

\item{type}{Either 2 (default) or 3 (note that either roman or arabic
numerals are acceptable). Type II tests each term after all other
non-higher-order terms, whereas type III tests each term after all other
terms, including higher-order terms.}

\item{contr_sum}{Logical. Refit model using sum-to-zero contrasts? Default is
\code{TRUE}.}
}
\value{
a \code{\link{data.frame}}
}
\description{
Performs F-test on coefficients of a model fit via glmmTMB using one
of several available methods for calculating denomenator degrees of freedom
}
