% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_disp}
\alias{extract_disp}
\title{Extract the possibly fixed residual dispersion parameter}
\usage{
extract_disp(mod, ...)
}
\arguments{
\item{mod}{A glmmTMB model.}

\item{...}{Ignored.}
}
\description{
This is a helper function used for getting \code{sigma()} when it is not estimated
but fixed.
}
