% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{power_contrast}
\alias{power_contrast}
\title{Power of Contrasts Performed on Models Fit with glmmTMB}
\usage{
power_contrast(emm, contr_list, alpha = 0.05, n_sims = 10000, ...)
}
\arguments{
\item{emm}{An emmGrid object associated with a \code{\link{glmmTMB}} model.}

\item{contr_list}{A (named) list of contrast specifications.}

\item{alpha}{Numeric. The nominal type I error rate. Defaults to 0.05.}

\item{n_sims}{Numeric. The number of simulations to use for calculating
type M error rate. If set to zero, the closed form, asymptotic
calculations are used.}

\item{...}{Other arguments passed to emmeans::contrast.}
}
\description{
This function calculates the power of contrasts.
}
\examples{
library(agridat)
library(dplyr)
library(glmmTMB)
library(emmeans)

# Power analysis in a split-plot RCBD experiment on oat yield
oats = agridat::yates.oats |> 
       dplyr::mutate(dplyr::across(nitro, factor))

# All random effects are arbitrarily given standard deviations of 10
oat_mod = glmmTMB(yield ~ gen*nitro + (1|block/gen), 
                  data = oats, 
                  REML = T, 
                  dispformula = ~0, 
                  start = list(theta = c(log(c(10, 10)))),
                  map = list(theta = factor(c(NA, NA))), 
                  control = glmmTMBControl(zerodisp_val = log(10))) 

# Here, the denominator DF are determined when EMMs are calculated:
(oat_emm = emmeans(oat_mod, ~ gen, ddf = 'kenward-roger'))

contr = list('GoldenRain vs Others' = c(1, -0.5, -0.5))

power_contrast(oat_emm, contr)             

# With manually specified denominator DFs (note that the argument
# is not `ddf =`, but `df =`)
(oat_emm2 = emmeans(oat_mod, ~ gen, df = 12))

power_contrast(oat_emm2, contr)  

}
