% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{power_contrast}
\alias{power_contrast}
\title{Power of Contrasts Performed on Models Fit with glmmTMB}
\usage{
power_contrast(emm, contr_list, ddf, alpha = 0.05)
}
\arguments{
\item{emm}{An emmGrid object associated with a \code{\link{glmmTMB}} model.}

\item{contr_list}{A named list of contrasts specifications.}

\item{ddf}{Numeric. The value to use for the denominator degrees of freedom.
Should probably correspond to the value returned from a call to
power_ftest.}

\item{alpha}{The nominal type I error rate. Defaults to 0.05.}
}
\description{
This function calculates the power of F-tests for each fixed
effect term in the model using one of several methods for determining the
denominator degrees of freedom |>
}
