% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{power_ftest}
\alias{power_ftest}
\title{Statistical Power of F-tests Performed on Models Fit with glmmTMB}
\usage{
power_ftest(mod, ddf = "containment", type = 2, alpha = 0.05)
}
\arguments{
\item{mod}{A \code{\link{glmmTMB}} model}

\item{ddf}{Method for calculating denominator degrees of freedom: containment
(default), pinheiro-bates, satterthwaite, or asymptotic.}

\item{type}{Either 2 (default) or 3 (note that roman or arabic numerals are
accepted). Type II tests each term after testing all other non-higher-order
terms, whereas type III tests each term after all others, including
higher-order terms.}

\item{alpha}{The nominal type I error rate. Defaults to 0.05.}
}
\description{
This function calculates the power of F-tests for each fixed
effect term in the model using one of several methods for determining the
denominator degrees of freedom |>
}
