% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.powertable}
\alias{print.powertable}
\title{Print the results of a power analysis}
\usage{
\method{print}{powertable}(x, digits = 1, pdigits = getOption("pdigits", default = 4), ...)
}
\arguments{
\item{x}{A \code{powertable} object}

\item{digits}{Integer. The number of digits to print for all columns other
than the p-value and power.}

\item{pdigits}{Integer. The number of digits to print for the p-value and
power columns}

\item{...}{unused}
}
\description{
This function controls the formatting when printing \code{powertable}
objects.
}
