% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_calcs.R
\name{theta_finder}
\alias{theta_finder}
\title{Identify how random effects terms will be ordered in the fitted glmmTMB model}
\usage{
theta_finder(formula, data, ...)
}
\arguments{
\item{formula}{A model formula (containing random effects).}

\item{data}{A data frame containing all the columns included in \code{formula}.}

\item{...}{Other arguments passed to \code{glmmTMB}.}
}
\description{
When performing a "plug-in" power analysis, the values for the random effects
must be supplied in the order in which they are saved internally in the glmmTMB
model. It is not always clear (especially where there are correlated random
slopes and intercepts) just what this ordering is. This helper function takes
a model formula and a data set and outputs a table enumerating all the implied random
effects and their ordering.
}
